TIMER: NORMAL & CTC modes

+++ TIMER_normalMode_outCompareUnit.c
@@ -19,7 +19,7 @@ int main()
 	while(1)
 	{
 		LCD_GoTo(0, 0);
-		LCD_WriteNumber((int)TIMER0_OCR0_READ());
+		LCD_WriteNumber(TIMER0_OCR0_READ());
 	}
 }
 
@@ -36,26 +36,26 @@ static void timer0CompareMatchFunction()
 //
 //void timer0CompareMatchISR ()
 //{
-//OCR0 = OCR0 + 10;
-//DIO_TogglePin(PINA0);
+	//OCR0 = OCR0 + 10;
+	//DIO_TogglePin(PINA0);
 //}
 //
 //int main()
 //{
-//DIO_Init();
-//TIMER0_Init ( TIMER0_PRESCALER_8, TIMER0_NORMAL, TIMER0_OC0_TOGGLE_ON_COMPARE_MATCH);
-//LCD_Init();
-//
-//SEI();
-//TIMER0_OCM_interruptEnable();
-//
-//TIMER0_compareMatchInterruptSet (timer0CompareMatchISR);
-//
-//OCR0 = 0;
-//
-//while (1)
-//{
-//LCD_GoTo(0,0);
-//LCD_WriteNumber_4Digit(OCR0);
-//}
+	//DIO_Init();
+	//TIMER0_Init ( TIMER0_PRESCALER_8, TIMER0_NORMAL, TIMER0_OC0_TOGGLE_ON_COMPARE_MATCH);
+	//LCD_Init();
+	//
+	//SEI();
+	//TIMER0_OCM_interruptEnable();
+	//
+	//TIMER0_compareMatchInterruptSet (timer0CompareMatchISR);
+	//
+	//OCR0 = 0;
+	//
+	//while (1)
+	//{
+		//LCD_GoTo(0,0);
+		//LCD_WriteNumber_4Digit(OCR0);
+	//}
 //}
